requirements.txt
fastapi
uvicorn
libsql
python-dotenv
pydantic
pydantic[email]